requirements.txt
Django==2.2
djangorestframework==3.9.0
djangorestframework-simplejwt==4.4.0

requests==2.19.1

flake8==3.7.7
pytest==4.0.1

Sphinx==1.8.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0

Unipath==1.1
urllib3==1.23

gunicorn==19.8.1
